Wait until mounted

+++ Pr2/FUSE_myFS/test1.sh
@@ -16,13 +16,15 @@ rm -R -f test
 mkdir test
 
 if [[ $automount = true ]]; then
-    pidof fs-fuse > /dev/null && { echo "fs-fuse ya se está ejecutando"; exit 1; }
+    grep -qs "$PWD/mount-point " /proc/mounts && { echo "Can't mount: Alredy mounted"; exit 1; } 
     [[ -d mount-point ]] && rm -rf mount-point
     mkdir mount-point
     ./fs-fuse -t 2097152 -a virtual-disk -f '-d -s mount-point' >fuse.log 2>fuserr.log &
     fusepid=$$
     echo "Started fuse with pid $fusepid"
-    sleep 1
+    while ! grep -qs "$PWD/mount-point " /proc/mounts; do
+        sleep 1
+    done
 fi
 
 echo 'file 1' > './test/file1.txt'
@@ -44,6 +46,6 @@ cat $MPOINT/file1.txt > './test/file1.txt.mpoint'
 diff -q 'test/file1.txt' './test/file1.txt.mpoint' || exit 1;
 
 if [[ $automount = true ]]; then
-    kill -s TERM -$fusepid
-    $! || { echo "Couldn't kill fuse"; cat fuserr.log; }
+    fusermount -u ./mount-point || { echo "Couldn't umount"; exit 1; }
+    # kill -s TERM -$fusepid || { echo "Couldn't kill fuse"; exit 1; }
 fi
